feat: add SomeVersionedEntity CRUD Playwright test with version assertions
Fixes Playwright Long serialization issue in assertVersionEquals by converting to Int before passing to waitForFunction.

+++ maia-showcase/app/src/test/kotlin/org/maiaframework/showcase/testing/pages/SomeVersionedBlotterPage.kt
@@ -142,7 +142,7 @@ class SomeVersionedBlotterPage(
             "  const c = document.querySelector('.ag-cell[col-id=\"version\"]');" +
             "  return c && c.innerText && c.innerText.trim() === String(v);" +
             "}",
-            expectedVersion
+            expectedVersion.toInt()
         )
     }
